+++ Sources/NablaScheduling/Views/Scenes/AppointmentConfirmation/AppointmentConfirmationViewModel.swift
@@ -281,7 +281,7 @@ final class AppointmentConfirmationViewModelImpl: AppointmentConfirmationViewMod
     
     private func price(for appointment: Appointment) -> String? {
         switch appointment.state {
-        case .finalized, .upcoming:
+        case .finalized, .scheduled:
             guard let price = appointment.price else { return nil }
             return format(price: price)
         case let .pending(paymentRequirement):